Fixing some tests.
